FlaskApplication/backend/requirements.txt
Flask==2.3.2
pymongo==4.5.0
python-dotenv==1.0.0
certifi
CORS